refactor phase 1 to reduce duplicate code
