+++ ReadRate/Views/Now Reading Screen/NowReadingList.swift
@@ -8,27 +8,6 @@
 
 import SwiftUI
 
-struct RoundedText: ViewModifier {
-    let style: Font.TextStyle
-    let bold: Bool
-    
-    func body(content: Content) -> some View {
-        if bold {
-            return content
-                .font(Font.system(style, design: .rounded).bold())
-        } else {
-            return content
-                .font(Font.system(style, design: .rounded))
-        }
-    }
-}
-
-extension Text {
-    func rounded(_ style: Font.TextStyle = .body, bold: Bool = true) -> some View {
-        self.modifier(RoundedText(style: style, bold: bold))
-    }
-}
-
 struct StyledButton: View {
     let iconName: String
     let label: String
